refactor: move controller definition to types
